pyproject.toml
[tool.poetry]
name = "inlinefragmentstrawberry"
version = "0.1.0"
description = ""
authors = ["Jeremy Berman <jerber@sas.upenn.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pip = "^22.3.1"
install = "^1.3.5"
strawberry-graphql = {extras = ["fastapi"], version = "^0.142.2"}
uvicorn = {extras = ["standard"], version = "^0.20.0"}
devtools = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
